test(mcp): move permission boundary to hono, skip CF serving suites
Always-proxying `/mcp` to Hono means the CF worker no longer serves MCP
via the durable object, so tests that exercised that serving path fail.

- Replace the worker → durable-object OAuth permission-error boundary
  test with an equivalent on Hono's `handleCatchError`, which now owns
  that boundary for all `/mcp` traffic (403 + insufficient_scope).
- Skip the CF real-stack serving suites (protocol, resilience, UI apps,
  resource catalog, tool behavior, catalog filtering) — the same
  behavior is covered against the Hono runtime that actually serves it.
  The bearer-token auth suite stays active since the worker still
  enforces 401s before proxying. Skipped, not removed, until the durable
  object is retired.

Generated-By: PostHog Code
Task-Id: 19a01b56-c02d-41f0-8532-2e6dc7c0eccb
